% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftMult-data.r
\docType{data}
\name{driftMult}
\alias{driftMult}
\title{Size bin and mesh size multiplers for comparing drift samples with differing mesh size.}
\format{
An object of class \code{"list"}.
}
\usage{
data(driftMultipliers)
}
\description{
Data on the proportion of total sample counts represented by each size bin for drift samples.
Also the multiplers used that vary from 250 to 500 um mesh nets.
Used internally for converting to/from 250 and 500 um mesh nets.
}
\examples{
data(driftMultipliers)
foo <- driftMult
}
\keyword{datasets}
