% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorGear.r
\name{errorGear}
\alias{errorGear}
\title{Check for errors in \code{gear} argument formatting.}
\usage{
errorGear(checkGear)
}
\arguments{
\item{checkGear}{The sampling gear type of interest (from \code{readDB} or \code{sampspec}).}
}
\value{
Runction proceeds invisibly if the gear type is OK. Otherwise it proceeds with a warning
if it can convert the gear type to something interpretable. Otherwise it returns an error.
}
\description{
Hidden function that makes sure the \code{gear} argument is formatted correctly and attempts to correct it if not.
}
\details{
Function checks for common errors in gear formatting such as lower case or misspellings.
It tries to correct these as best as possible.
}
\examples{
## Correct a gear misspelling (lowercase)
foo <- readDB(gear = "drift")
}
\seealso{
\code{\link{readDB}} and \code{\link{sampspec}}, which utilize this function.
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}
}
\concept{access, database}
\keyword{internal}
