% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodbase.r
\name{foodbase}
\alias{foodbase}
\title{Read, format, and combine data from the Foodbase database}
\usage{
foodbase(gear = "Drift", path = "Network", samp = "", spec = "",
  sppl = "", species = "All")
}
\arguments{
\item{gear}{The sampling gear type of interest (\code{Drift}, \code{LightTrap}, etc). Default is \code{Drift}.}

\item{path}{The download location of the data on the \code{Network}, on \code{GitHub}, or elsewhere. Default is \code{Network}.}

\item{samp}{The name of the sample dataframe or the location of the \code{Sample.csv} file. See details.}

\item{spec}{The name of the specimen dataframe or the location of the \code{Specimen.csv} file. See details.}

\item{sppl}{The name of the species list dataframe or the location of the \code{SppList.csv} file. See details.}
}
\value{
Creates a list containing the following dataframes:\cr
\code{Samp}: The sample data.\cr
\code{Spec}: The specimen data.\cr
\code{Stats}: Statistics (counts, richness, sizes, etc.) on the specimen data, by sample.\cr
\code{Counts}: An ordination-friendly matrix of count data, by taxon, by sample.\cr
\code{SizeMeans}: An ordination-friendly matrix of mean sizes, by taxon, by sample.\cr
\code{SizeMedians}: An ordination-friendly matrix of median sizes, by taxon, by sample.\cr
\code{SizeSDs}: An ordination-friendly matrix of the standard deviation in sizes, by taxon, by sample.\cr
\code{Biomass}: An ordination-friendly matrix of biomass, by taxon, by sample.\cr
\code{Spp}: The taxa list of taxa in the dataset.\cr
\code{Missing}: Sample data for any samples that don't have corresponding specimen data (or the species of interest) and were cut from the dataframes listed above.\cr
\code{SampDel}: Sample data for any samples that were in the read sample data but were flagged for deletion and therefore cut from the dataframes listed above.\cr
\code{SpecDel}: Specimen data for any samples that were in the read sample data but were flagged for deletion and therefore cut from the dataframes listed above.

Note on units: All biomass values are in \code{mg}, and sizes are in \code{mm}. \code{Distance} is in \code{m}, \code{Velocity} is in \code{m/s} and \code{Volume} is in \code{m^3/s}. \code{TimeElapsed} is in \code{seconds} and \code{ProcessTime} is in decimal \code{hours}.
}
\description{
Pulls exported data from the Foodbase database for use in R, combines Sample, Specimen, and Species code data, formats the data in a variety of ways to facilitate analysis, and runs some simple statistics.
}
\details{
Currently only \code{Drift} is implemented for \code{gear}.

The \code{path} argument is helpful if you are off the USGS network, in which case data can be downloaded from the web by specifying \code{path = "GitHub"}. Note that data are not routinely pushed to GitHub, however, so the data there may be slightly dated. Alternately, you can specify \code{path} to a local directory (see examples).

The function will look for \code{samp}, \code{spec}, and \code{sppl} within the directory of \code{path}, unless these are specified individually. Any entry for \code{samp}, \code{spec}, and \code{sppl} overrides \code{path}. See examples. 

In general, it is best to use the companion function \code{\link{readDB}} first to get sample data, subset those sample data, then run \code{foodbase} using those sample data as \code{samp}.
}
\examples{
## Read in drift sample data from the network (all the defaults).
foo <- readDB()

## Subset only data from Lees Ferry
foo2 <- foo[foo$Reach == "CRLeesFerry",]

## Get the specimen data for these samples, all wrapped together and formatted nicely.
foo3 <- foodbase(samp = foo2)

## Or, if you want to analyze all drift data in the database for inexplicable some reason:
whyme <- foodbase(gear = "Drift")

## Example of all data stored in a local directory.
loc <- foodbase(path = "C:/Users/nflanders/Documents/Analysis/")

## Example of sample data stored in a local directory, while species data are not.
mysamp <- read.csv("C:/Users/nflanders/Documents/Analysis/DriftSample.csv")
loc <- foodbase(path = "Network", samp = mysamp)

## Example for downloading data from GitHub when off the USGS network.
gith <- foodbase(path = "GitHub")

## Example to get only samples with New Zealand mudsnails.
nzms <- foodbase(species = "NZMS")
}
\seealso{
\code{\link{readDB}}, for simply reading in the Sample or Specimen data from the Foodbase database individually, with no fanfare.
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}
}
\concept{
access, database
}
