% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordmat.r
\name{ordmat}
\alias{ordmat}
\title{Create a matrix of specimen data}
\usage{
ordmat(sampspec, type = "count")
}
\arguments{
\item{sampspec}{The name of the \code{sampspec} list output containing sample and specimen information for building the matrix.}

\item{type}{The type of matrix output desired. See details. The default, \code{count}, creates a matrix of count data by taxon and sample.}
}
\value{
Creates a dataframe where each row is a sample and each column is a taxon.
}
\description{
Takes data from sampspec and builds ordination-friendly matrices of counts, mean/median/SD sizes, or biomasses, by taxon and sample.
}
\details{
This function will likely only work with \code{sampspec} set to an output generated by the companion \code{\link{sampspec}} function. If you are working with data that do not have measured sizes and only counts (e.g., light trap data), then biomass and size matrices cannot be made and will return an error.

The \code{type} argument can be specified to return a matrix of counts (\code{count}), biomass (\code{biomass}), or mean, median, or standard deviation in sizes (\code{sizemean}, \code{sizemed}, \code{sizesd}, respectively).

Note on units: All biomass values are in \code{mg}, and sizes are in \code{mm}.
}
\examples{
## Read in drift sample data from the network (all the defaults).
foo <- readDB()

## Subset only data from Lees Ferry
foo2 <- foo[foo$Reach == "CRLeesFerry",]

## Get the specimen data for these samples, all wrapped together and formatted nicely.
foo3 <- sampspec(samp = foo2)

## Create an ordination matrix of biomass data
myout <- ordmat(sampspec = foo3, type = "biomass")
}
\seealso{
\code{\link{sampspec}}, for creating the input list required for making \code{ordmat} work.
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}
}
\concept{
access, database, ordination
}
