% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packload.r
\name{packload}
\alias{packload}
\title{Install, update, and load packages}
\usage{
packload(packages, updater = FALSE, quiet = TRUE)
}
\arguments{
\item{packages}{The names of the R packages to be loaded. Specified 
with quotes.}

\item{updater}{Whether to search for and istall any updates to packages,
or just load existing packages (assuming they are already installed).
Any packages specified that are not already installed will be 
installed with the latest versions regardless. Default is \code{FALSE}.}

\item{quiet}{Whether to load packages with as little commentary as 
possible. Default is \code{TRUE}.}
}
\value{
Loads the packages of interest, for use in the R workspace.
}
\description{
Function to check if specified R packages are currently 
installed on the user's machine. It installs them if not, updates 
them if selected, and then loads them.
}
\details{
The function checks for packages located on CRAN and on Jeff Muehlbauer's
GitHub space (\link{https://github.com/jmuehlbauer-usgs/R-packages}). 
In general specifying \code{updater = TRUE} is unnecessary unless you know a big 
an R package has recently been developed. The \code{quiet} argument tries its
best to prevent R's printing of package loading commentary, but some text may 
still get through.
}
\examples{
# Load Jeff's "bugR" package and the CRAN package "lubridate".
packload(c("bugR", "lubridate"))
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}
}
\concept{packages}
