% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampspec.r
\name{sampspec}
\alias{sampspec}
\title{Read, format, and combine data from the Foodbase database}
\usage{
sampspec(
  samp = "",
  spec = "",
  sppl = "",
  species = "All",
  stats = FALSE,
  gear = "",
  mesh = "origin"
)
}
\arguments{
\item{samp}{The name of the sample dataframe, if not working from
auto-downloaded data. See Details.}

\item{spec}{The name of the specimen dataframe, if not working from
auto-downloaded data. See Details.}

\item{sppl}{The name of the species list dataframe, if not working from
auto-downloaded data. See Details.}

\item{species}{Whether to subset the data for only a given taxon. See
Details. Default is \code{"All"}.}

\item{stats}{Whether to calculate total count, size, and biomass data for
each taxon in each sample. Default is \code{FALSE}.}

\item{gear}{The sampling gear type of interest (\code{Drift},
\code{LightTrap}, \code{Sticky}, or \code{FishGut}). Should be specified
only in rare cases where you are not working from \code{\link{readDB}}
output. See Details.}

\item{mesh}{The desired mesh size to be used for analysis of drift data
(\code{"origin"}, \code{"500"}, or \code{"250"}). See Details.
Default is \code{"origin"}.}
}
\value{
Creates a list containing the following dataframes:\cr
  \code{Samples}: The sample data.\cr
  \code{Specimens}: The specimen size
  data. Accounts for Count Extra counts by assigning them proportionally to
  size bins.\cr
  \code{Biomass}: The specimen biomass data. Accounts for Count
  Extra counts by assigning them proportionally to size bins.\cr
  \code{RawSpecimens}: The specimen size data, excluding any Count Extra
  counts.\cr
  \code{RawBiomass}: The specimen biomass data, excluding any
  Count Extra counts.\cr
  \code{Taxa}: The taxa list of taxa in the
  dataset.\cr
  \code{Missing}: Sample data for any samples that don't have
  corresponding specimen data (or the species of interest) and were cut from
  the dataframes listed above.\cr
  \code{SampDel}: Sample data for any samples
  that were in the read sample data but were flagged for deletion and
  therefore cut from the dataframes listed above.\cr
  \code{SpecDel}: Specimen
  data for any samples that were in the read sample data but were flagged for
  deletion and therefore cut from the dataframes listed above.\cr
  \code{Statistics}: Total count, size, and biomass data, by specimen (if
  \code{stats = TRUE}).

Note on units: All count data are presented as raw counts (i.e., just number
of bugs, and not density, rate, or concentration). All biomass values are in
\code{mg} except for the mass of fish whose guts are being analyzed, which are
in \code{g}. All sizes are in \code{mm}. \code{Distance} is in \code{m},
all depths are in \code{ft}, \code{Velocity} is in \code{m/s} and 
\code{Volume} is in \code{m^3}. 
\code{TimeElapsed} is in decmial \code{minutes} for Drift, decimal 
\code{hours} for Light Traps, and decimal \code{days} for Sticky traps.
\code{ProcessTime} is in decimal
\code{hours}. \code{RiverMile} is generally in \code{miles} unless otherwise
specified in the \code{Notes}.

Note on compatibility: If you plan to use \code{\link{sampstats}} or
\code{\link{ordmat}} on the \code{sampspec} output, then set \code{stats =
TRUE}.
}
\description{
Pulls exported data from the Foodbase database for use in R,
  combines Sample, Specimen, and Species List data, and formats the data to
  facilitate analysis.
}
\details{
The data are based on data saved locally on your computer from the Foodbase
database when you run the function \code{\link{readDB}}. To update these
data, use \code{\link{readDB}} (see Examples).

The function will look for \code{samp}, \code{spec}, and \code{sppl} within
the \code{Data/} directory of \code{foodbase}, unless these are specified
individually. Any entry for \code{samp}, \code{spec}, and \code{sppl}
overrides data in the \code{Data/} directory of \code{foodbase} (see
Examples).

In general, it is best to use the companion function \code{\link{readDB}}
first to get sample data, subset those sample data, then run \code{sampspec}
using those sample data as \code{samp}.

Using \code{species} you can return data for only certain taxa of interest.
In addition to choosing species individually (e.g., \code{species = "CHIL"}
or \code{species = c("CHIL", "SIML")}), you can also use the shortcut
\code{species = "Big4"} to subset only for species codes \code{CHIL},
\code{SIML}, \code{NZMS}, and \code{GAMM}, the shortcut \code{species =
"Big9"} to subset only for species codes \code{CHIL}, \code{CHIP},
\code{CHIA}, \code{SIML}, \code{SIMP}, \code{SIMA}, \code{OLIG}, \code{NZMS},
and \code{GAMM}, or the shortcut \code{species = "Caddis"} to subset only 
for caddisfly species codes.

The argument \code{gear} can be specified (e.g., \code{gear = 'Drift'}), and
in general there is no harm in doing so. However, in most cases \code{gear}
will inherit the sample type from the attributes of the \code{\link{readDB}}
output, so specifying it here is unnecessary. The exception is in rare cases
where \code{\link{readDB}} is not run before running \code{sampspec}, as in
the case of the \code{whyme} example below.

The argument \code{mesh} only applies to Drift data (\code{gear = 'Drift'}).
It allows apples-to-apples comparison of samples collected using 250 and 500
um mesh nets, based on a series of bin-specific conversion multipliers.
These multipliers are stored as a list of dataframes; see \code{\link{driftMult}}
for more information. The default (\code{"origin"}) does no conversion of
specimen data, while specifying \code{mesh = "500"} will convert all specimen
data from samples collected using 250 um mesh nets/sieves and return data on 
the scale of 500 um data, using the mutliplers mentioned above (and vice versa
for \code{mesh = "250"}). Generally, converting from the finer 250-um to the 
coarser 500-um mesh is the preferred direction for accuracy, due to 
zero-boundedness in the data. These conversions apply only to \code{Specimens}, 
\code{Biomass}, and \code{Statistics} (if specified) dataframes in the output,
and only for Drift data. Other dataframes in the output (e.g., 
\code{RawSpecimens}) retain the original, unconverted counts.
}
\examples{
# Read in drift sample data from the network, and update sample, specimen, and species list data.
foo <- readDB(updater = TRUE)

# Subset only data from Lees Ferry
foo2 <- foo[foo$Reach == "CRLeesFerry",]

# Get the specimen data for these samples, all wrapped together and formatted nicely.
foo3 <- sampspec(samp = foo2)

# Or, if you want to analyze all drift data in the database for some inexplicable reason:
whyme <- sampspec(gear = "Drift")

# Example to get only drift samples with New Zealand mudsnails.
nzms <- sampspec(species = "NZMS", gear = "Drift")
}
\seealso{
\code{\link{readDB}}, for initial read in of the Sample or Specimen
  data from the Foodbase database individually, and updating these data.
  \code{\link{ordmat}} for creating ordination-type matrices of counts,
  sizes, or biomasses by sample and taxon from a \code{sampspec} output.
  \code{\link{sampstats}} for computing sample-level statistics on the data.
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov} and Michael J. Dodrill, \email{mdodrill@usgs.gov}
}
\concept{access, database}
