% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampstats.r
\name{sampstats}
\alias{sampstats}
\title{Sample-level statistics on foodbase data}
\usage{
sampstats(sampspec)
}
\arguments{
\item{sampspec}{Output from the \code{\link{sampspec}} function. See Details.}
}
\value{
Creates a dataframe containing totaled counts, richness, mean/median/standard deviation in size, and biomass for each sample, as well as the same values for aquatic invertebrates, aquatic insects, EPT taxa, and aquatic Diptera.
}
\description{
Runs some simple statistics on formatted data sets from the Foodbase database.
}
\details{
Currently only implemented for Drift samples.

This function will likely only work with \code{sampspec} set to an output generated by the companion \code{\link{sampspec}} function, and with \code{sampspec(..., stats = TRUE)}. See Examples.
Note on units: All count data are presented as raw counts (i.e., just number of bugs, and not density, rate, or concentration). All biomass values are in \code{mg}, and sizes are in \code{mm}.
}
\examples{
## Read in drift sample data from the network (all the defaults).
foo <- readDB()

## Subset only data from Lees Ferry
foo2 <- foo[foo$Reach == "CRLeesFerry",]

## Get the specimen data for these samples, all wrapped together and formatted nicely.
foo3 <- sampspec(samp = foo2)

## Get statistics on these data
myout <- sampstats(foo3)
}
\seealso{
\code{\link{sampspec}}, for creating the input list required for making \code{sampstats} work.
}
\author{
Jeffrey D. Muehlbauer, \email{jmuehlbauer@usgs.gov}
}
\concept{
access, database
}
